Move common dataset/composite filters to dataobject filters
